+++ src/main/java/com/calendar/repository/UserJpaRepository.java
@@ -21,7 +21,7 @@ public interface UserJpaRepository extends JpaRepository<User, Long> {
 			"concat(u.a_pat,' ',u.a_mat,' ',u.nombre) as nombreProfesional\n" + 
 			"from users u, profesional p\n" + 
 			"where u.perfil = 1 and u.id = p.fk_idUsuario and p.cod_centro=:idCentro", nativeQuery = true)
-	List<Map<String, Object>> getUserProfesionals(@Param("idCentro") int idCentro);
+	List<Map<String, Object>> getUserProfesionals(@Param("idCentro") Long idCentro);
 
 	public abstract User findUserByEmailAndVigente(String email, int vigente);
