+++ src/main/java/com/calendar/service/UserService.java
@@ -12,7 +12,7 @@ public interface UserService {
 	
 	public abstract User findUserByEmail(String email);
 	public abstract User findByEmailAndPass(String email, String pass);
-	public abstract List<Map<String, Object>> getUserProfesionals(int idCentro);
+	public abstract List<Map<String, Object>> getUserProfesionals(Long idCentro);
 	public abstract User findUserByEmailAndVigente(String email, int vigente);
 	List<Map<String, Object>> getListaUsuariosCentro(Long idCentro);
